🐛 Ignore stderr from erblint
Null LS expects JSON output from erblint even for the stderr messages.

When erblint is run in a project without a config file it outputs a
message on stderr in plain text that states that it will use the default
config since no config could be found.

This causes Null LS to stop executing erblint since it could not parse
the error message as JSON. And, because of this, Null LS will not
execute erblint after that point.

I want erblint to run with it's default linting settings in all
projects even if it hasn't been configured.